Clarify "if we can" and incorporate CASK
Clarify that respect isn't an "if you can" optional — I think the original intent here was "if you can _answer the question_" but I think this is clearer. 

Also: IMHO sometimes good programmers, colleagues, teachers do have to judge work (not people, but work) — in order to help provide feedback and make the work better / help their collaborators grow. But should we be CASK in giving that feedback? Yes!